+++ src/test/java/com/epam/reportportal/junit5/features/parameters/EnumParametersFailedTest.java
@@ -1,12 +1,13 @@
 package com.epam.reportportal.junit5.features.parameters;
 
 import com.epam.reportportal.junit5.ParametersTest;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
 @ExtendWith(ParametersTest.ParameterTestExtension.class)
-public class EnumParametersTest {
+public class EnumParametersFailedTest {
 
 	public enum TestParams {
 		ONE,
@@ -17,6 +18,9 @@ public class EnumParametersTest {
 	@EnumSource(TestParams.class)
 	public void testParameters(TestParams param) {
 		System.out.println("Test: " + param.name());
+		if (param == TestParams.ONE) {
+			Assertions.fail("Failure parameter passed: " + param);
+		}
 	}
 
 }